Extract card details component
