+++ src/components/game/DrawnCard.css
@@ -12,49 +12,3 @@
   grid-template-columns: auto;
   grid-template-rows: auto;
 }
-
-.drawnCardWrapper {
-  grid-column: 1;
-  grid-row: 1;
-  place-self: center;
-}
-
-.drawnCard {
-  min-width: 300px;
-  max-width: 400px;
-  border-radius: 25px;
-  border: 3px solid black;
-}
-
-.cardContentsContainer {
-  padding: 20px;
-}
-
-.cardName {
-  text-align: center;
-  font-size: 28px;
-}
-
-.cardSubtype {
-  text-align: center;
-}
-
-.cardFlavorText {
-  font-style: italic;
-  text-align: center;
-}
-
-.rollTable {
-  border-spacing: 25px 10px;
-}
-
-.drawnCardControls {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  padding: 20px;
-}
-
-.drawnCardControls > * {
-  height: 50px;
-}
